Factor out parse_attr_names
